+++ ICD.Common.Utils.Tests/Json/JsonUtilsTest.cs
@@ -10,29 +10,7 @@ namespace ICD.Common.Utils.Tests.Json
 	public sealed class JsonUtilsTest
 	{
 		[Test]
-		public void CacheTypeTest()
-		{
-			Assert.Inconclusive();
-		}
-
-		[Test]
-		public void ParseDateTimeTest()
-		{
-			const string dataA = "2016-02-26T19:24:59";
-			const string dataB = "2019-04-01T12:41:15-04:00";
-
-			Assert.DoesNotThrow(() => JsonUtils.ParseDateTime(dataA));
-			Assert.DoesNotThrow(() => JsonUtils.ParseDateTime(dataB));
-		}
-
-		[Test]
-		public void TryParseDateTimeTest()
-		{
-			Assert.Inconclusive();
-		}
-
-		[Test]
-		public void PrintTest()
+		public void FormatTest()
 		{
 			Assert.Inconclusive();
 		}