+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - Fixed a bug where ANSI color encoded strings with percentages were being scrambled
+ - Improvements to JSON DateTime parsing, particularly in Net Standard
 
 ## [9.7.0] - 2019-08-15
 ### Added